requirements.txt
flask
gunicorn
pytest
pyyaml